Create log directory when accessed
